+++ tests/test_proposal.py
@@ -14,7 +14,7 @@ def test_get_admin(mock_user, client):
 @pytest.mark.parametrize("mock_user", [admin], indirect=True)
 def test_search_code(mock_user, client):
     """Get all proposals with a matching proposal code"""
-    resp = client.get("/proposals?s=bi")
+    resp = client.get("/proposals?search=bi")
     assert resp.status_code == 200
     assert resp.json()["total"] == 1
 
@@ -22,7 +22,7 @@ def test_search_code(mock_user, client):
 @pytest.mark.parametrize("mock_user", [admin], indirect=True)
 def test_search_number(mock_user, client):
     """Get all proposals with a matching proposal number"""
-    resp = client.get("/proposals?s=1")
+    resp = client.get("/proposals?search=1")
     assert resp.status_code == 200
     assert resp.json()["total"] == 3
 
@@ -30,7 +30,7 @@ def test_search_number(mock_user, client):
 @pytest.mark.parametrize("mock_user", [admin], indirect=True)
 def test_search_full(mock_user, client):
     """Get all proposals with a matching proposal code and number"""
-    resp = client.get("/proposals?s=cm31")
+    resp = client.get("/proposals?search=cm31")
     assert resp.status_code == 200
     assert resp.json()["total"] == 1